backend/requirements.txt
fastapi
uvicorn[standard]
celery
redis
biopython
numpy
python-multipart